backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9
pandas==2.2.0
openai==1.55.0
anthropic==0.40.0
python-dotenv==1.0.0
pydantic==2.9.0
